Refactor GPU executable implementation
