+++ xla/service/gpu/BUILD
@@ -695,13 +695,58 @@ bool_flag(
 )
 
 cc_library(
+    name = "command_buffer_va_remapping",
+    srcs = ["command_buffer_va_remapping.cc"],
+    hdrs = ["command_buffer_va_remapping.h"],
+    deps = [
+        ":buffer_allocations",
+        "//xla:status_macros",
+        "//xla:util",
+        "//xla/backends/gpu/runtime:command",
+        "//xla/backends/gpu/runtime:command_buffer_thunk",
+        "//xla/backends/gpu/runtime:thunk",
+        "//xla/backends/gpu/runtime:thunk_executor",
+        "//xla/runtime:buffer_use",
+        "//xla/service:buffer_assignment",
+        "//xla/service:hlo_proto_cc",
+        "//xla/service:logical_buffer",
+        "//xla/service:service_executable_run_options",
+        "//xla/stream_executor:device_address",
+        "//xla/stream_executor:device_address_allocator",
+        "//xla/stream_executor:memory_allocation",
+        "//xla/stream_executor:memory_reservation",
+        "//xla/stream_executor:stream",
+        "//xla/stream_executor:stream_executor_h",
+        "//xla/stream_executor:vmm_device_address_allocator",
+        "@com_google_absl//absl/algorithm:container",
+        "@com_google_absl//absl/base:core_headers",
+        "@com_google_absl//absl/container:btree",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/container:node_hash_map",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings:str_format",
+        "@com_google_absl//absl/strings:string_view",
+        "@com_google_absl//absl/synchronization",
+        "@com_google_absl//absl/types:span",
+    ],
+)
+
+cc_library(
     name = "gpu_executable",
-    srcs = ["gpu_executable.cc"],
+    srcs = [
+        "gpu_executable.cc",
+        "gpu_executable_serialization.cc",
+        "gpu_executable_thunk_execution.cc",
+    ],
     hdrs = ["gpu_executable.h"],
     deps = [
         ":alias_info",
         ":backend_configs_cc",
         ":buffer_allocations",
+        ":command_buffer_va_remapping",
         ":gpu_constants",
         ":gpu_executable_proto_cc",
         ":gpu_executable_run_options",
